Add a bunch of failing tests
Some tests of functionality that should work but for
whatever reason doesn't, were added.

They include application in the style of
fun(X) -> math:log10(X) end(10)

with partially applied functions as
math:log10(_)(10)

which should work to be consistent with the rest of
Erlang, even if it makes little sense to actually do so.